chore(frontend/api-response-types): added types for type safe api response usage
