--- frontend/types/api-endpoints-types.ts
@@ -1,1 +1,0 @@
-export type Endpoint = "home" | "properties" | "featured";
